move prettytable into inc.common
